wecp initial commit

+++ backend/src/main/java/com/wecp/progressive/repository/LoanRepository.java
@@ -1,11 +1,10 @@
 package com.wecp.progressive.repository;
 
+
+import com.wecp.progressive.entity.Loan;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.wecp.progressive.entity.Loan;
-
 @Repository
-public interface LoanRepository extends JpaRepository<Loan, Integer>{
-    
+public interface LoanRepository extends JpaRepository<Loan, Long> {
 }
